interview_agent/pyproject.toml
[project]
name = "interview-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.133.1",
    "langchain>=1.2.10",
    "langchain-groq>=1.1.2",
    "langgraph>=1.0.9",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "uvicorn>=0.41.0",
]


